Task/33 store models in database (#37)
While models can be generated from LDR files they was not prior way to
store a model object sides writing it to disk as a file. Now we can
store models in the `models` table in `postgres`. The only current issue
is that to store gpb we store the model gpb data as a `byte[]` but to be
able to support easy look ups for postgres queryable data also needs to
be written to columns as well.