[Task] - Switch using protobuf binary data to reference / lookup tables in Postgres (#45)
Model data was being stored in the database as a `bytea` type. This was
not going to work long term as data wouldn't be migratable as they need
to be in primitive types to easily update later. Switched to seperating
out parts of a `Model` into seperate tables that can be joined later